data_ingestion_service/app/requirements.txt
fastapi
uvicorn[standard]
pydantic
httpx
pandas
minio
python-multipart # For file uploads
python-dotenv
# For in-memory CSV creation
io